petstagram/requirements.txt
asgiref==3.11.0
Django==6.0.1
django-unfold==0.76.0
pillow==12.1.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
